# api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@trexis.net
    name: treXis LLC
    url: https://www.trexis.net
  description: The public facing API through which connectors are exposed as a single
    abstract API
  termsOfService: http://www.trexis.net/termsofservices
  title: Cloud API
  version: v1.5
servers:
- url: /
security:
- cloudAPI: []
- connectorToken: []
paths:
...
  /cloud/v1/account/{accountId}/cards:
    get:
      description: Return the list of cards of a given account
      operationId: getAccountCards
      parameters:
      - description: Unique Account Identifier
        explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      - description: Indicates if data in the response should be masked
        explode: true
        in: query
        name: mask
        required: false
        schema:
          default: true
          type: boolean
        style: form
      - description: Indicates if the request should be processed for enhancement
          after getting data from the connector
        explode: true
        in: query
        name: enhance
        required: false
        schema:
          default: false
          type: boolean
        style: form
      - description: Value used for tracing requests and responses
        explode: false
        in: header
        name: X-TRACE-ID
        required: false
        schema:
          example: "1234"
          type: string
        style: simple
      - description: Token to carry/send information between requests. Connectors
          generate and validate the token content and type.
        explode: false
        in: header
        name: X-TOKEN
        required: false
        schema:
          example: VXNlclNlc3Npb25Ub2tlbjoxMjM0MTIzNA==
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardsList'
          description: Cards of the requested account
          headers:
            X-TRACE-ID:
              $ref: '#/components/headers/X-TRACE-ID'
            X-TOKEN:
              $ref: '#/components/headers/X-TOKEN'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Return a accounts cards
      tags:
      - Cards
...
components:
  headers:
    X-TRACE-ID:
      description: Value used for tracing requests and responses
      explode: false
      required: false
      schema:
        example: "1234"
        type: string
      style: simple
    X-TOKEN:
      description: Token to carry/send information between requests. Connectors generate
        and validate the token content and type.
      explode: false
      required: false
      schema:
        example: VXNlclNlc3Npb25Ub2tlbjoxMjM0MTIzNA==
        type: string
      style: simple
...
  schemas:
...
    Card:
      description: Card object
      properties:
        type:
          description: The type of card
          enum:
          - CREDIT
          - DEBIT
          - LOYALTY
          type: string
        id:
          description: Unique card identifier in the system of record
          example: 12345ABC
          type: string
        number:
          description: Card number (masked if masked=true)
          example: xxxx-xxxx-xxxx-1234
          type: string
        startDate:
          description: ISO 6801 date of when the card is started/activated
          example: YYYY-MM-DDThh:mm:ss
          type: string
        expiryDate:
          description: ISO 6801 date of when the card expires
          example: YYYY-MM-DDThh:mm:ss
          type: string
        holderName:
          description: The name of the cardholder
          example: Dr. Luke Skywalker
          type: string
        network:
          description: The card network program
          example: Mastercard
          type: string
        status:
          description: The status of the card
          example: LOCKED
          type: string
        URI:
          $ref: '#/components/schemas/FiniteURI'
      type: object
...
    FiniteURI:
      description: Unique Resource Identifier in the API
      example:
        URL: http://host/images/1234.jpg
      properties:
        URL:
          description: The URL to the URI
          example: http://host/images/1234.jpg
          type: string
      type: object
...
    BaseListProperties:
      description: Base properties for all lists object
      properties:
        status:
          description: Indicates whether a operation was successful
          type: boolean
        messages:
          description: Potential information or errror messages
          items:
            description: Potential information or error messages
            type: string
          type: array
        totalItems:
          description: The total number of list items that exist for the list or query
          example: 200
          type: integer
        nextCursor:
          description: The Base64 encoded value that indicates the next cursor. The
            cursor will be decoded to use as the starting point to fetch the next
            set of results.
          example: YWNjb3VudE51bWJlcjoxMjM0MTIzNCxhY2NvdW50VHlwZT1EREE=
          type: string
      type: object
...
    CardsList:
      allOf:
      - $ref: '#/components/schemas/BaseListProperties'
      - $ref: '#/components/schemas/CardsList_allOf'
...
    BaseErrorProperties:
      description: Base properties for all lists object
      properties:
        status:
          description: Indicates whether a operation was successful
          type: boolean
        messages:
          description: Potential information or errror messages
          items:
            description: Potential information or error messages
            type: string
          type: array
      type: object
    ErrorResponse:
      allOf:
      - $ref: '#/components/schemas/BaseErrorProperties'
...
    CardsList_allOf:
      properties:
        cards:
          description: The list of cards
          items:
            $ref: '#/components/schemas/Card'
          type: array
...
  securitySchemes:
    cloudAPI:
      description: The unique tenant API key allowing access to the API and connectors
      in: header
      name: X-API-KEY
      type: apiKey
    connectorToken:
      description: The unique key that can be passed to the connector
      in: header
      name: X-TOKEN
      type: apiKey